+++ RELEASES.md
@@ -19,6 +19,7 @@ New features
 
 * `OptimizationLevel::Simple` now eagerly evaluates built-in binary operators of primary types (if not overloaded).
 * Added `is_def_var()` to detect if variable is defined, and `is_def_fn()` to detect if script function is defined.
+* `Dynamic::from(&str)` now constructs a `Dynamic` with a copy of the string as value.
 
 
 Version 0.19.0